\name{dot-PBSmapEnv}
\docType{data}
\alias{.PBSmapEnv}
\title{PBSmapping Environment}
\description{
  An environment set aside for PBSmapping.
}
\usage{.PBSmapEnv}
\format{
A new environment with a \code{.GlobalEnv} parent.
}
\source{
  Generated by a call to the base function \code{new.env()}.
}
\details{
  The environment is created in \code{'zzz.r'} and can be
  used by \code{PBSmodelling} functions \code{'lisp'}, \code{'tget'}, 
  \code{'tput'}, \code{'tprint'}, and \code{'tcall'}, if this package has be loaded.
  
  Currently, \code{'.PBSmapEnv'} is only used in \code{'.validateData'}
  (in the namespace), and by a set of \code{print} functions specific to PBSmapping.
}
\seealso{
In \pkg{PBSmapping}:\cr
  \code{\link[PBSmapping]{print.EventData}},
  \code{\link[PBSmapping]{print.LocationSet}},
  \code{\link[PBSmapping]{print.PolyData}},
  \code{\link[PBSmapping]{print.PolySet}},
  \code{\link[PBSmapping]{print.summary.PBS}}\cr
In \pkg{PBSmodelling}:\cr
  \code{lisp},
  \code{tget},
  \code{packList}
}
\keyword{environment}

