\name{importLocs}
\alias{importLocs}
\title{Import LocationSet from a text file}
\description{
  Import a text file and convert into a LocationSet.
}
\usage{
importLocs(LocationSet)
}
\arguments{
  \item{LocationSet}{filename of LocationSet text file.}
}
\value{
  An imported LocationSet.
}
\author{
  \href{https://academic.macewan.ca/boersn/}{Nicholas M. Boers}, Associate Professor -- Computer Science\cr
  MacEwan University, Edmonton AB\cr
  Last modified \code{Rd: 2007-06-06}
}
\seealso{
  \code{\link{importPolys}}, \code{\link{importEvents}}, \code{\link{importGSHHS}}, \code{\link{importShapefile}}
}
\keyword{file}
