\name{PBSmapping}
\alias{PBSmapping}
\alias{PBSmapping-package}
\title{PBS Mapping: Draw Maps and Implement Other GIS Procedures}
\description{
  This software has evolved from fisheries research
  conducted at the Pacific Biological Station (PBS) in Nanaimo,
  British Columbia, Canada. It extends the R language to include
  two-dimensional plotting features similar to those commonly
  available in a Geographic Information System (GIS).  Embedded
  C code speeds algorithms from computational geometry, such as
  finding polygons that contain specified point events or
  converting between longitude-latitude and Universal Transverse
  Mercator (UTM) coordinates.  It includes data for a global
  shoreline and other data sets in the public domain.

  For a complete user's guide, see the file \code{PBSmapping-UG.pdf} 
  in the R directory \code{.../library/PBSmapping/doc}.

  \code{PBSmapping} includes 10 demos that appear as figures in 
  the User's Guide. To see them, run the function \code{.PBSfigs()}.

  More generally, a user can view all demos available from locally 
  installed packages with the function \code{runDemos()} in our 
  related (and recommended) package \code{PBSmodelling}.
}
\author{
  \href{https://github.com/boersn/}{Nicholas M. Boers}, Staff Software Engineer\cr
  Jobber, Edmonton AB\cr
  Last modified \code{Rd: 2008-09-03}
}
\keyword{documentation}
