\name{importLocs}
\alias{importLocs}
\title{Import LocationSet from a text file}
\description{
  Import a text file and convert into a LocationSet.
}
\usage{
importLocs(LocationSet)
}
\arguments{
  \item{LocationSet}{filename of LocationSet text file.}
}
\value{
  An imported LocationSet.
}
\seealso{
  \code{\link{importPolys}}, \code{\link{importEvents}}, \code{\link{importGSHHS}}, \code{\link{importShapefile}}
}
\keyword{file}
