\name{importLocs}
\alias{importLocs}
\title{Import LocationSet from Text File}
\description{
  Import a text file and convert it into a LocationSet.
}
\usage{
importLocs(LocationSet)
}
\arguments{
  \item{LocationSet}{\code{character} -- file name of LocationSet text file.}
}
\value{
  An imported LocationSet.
}
\author{
  \href{https://github.com/boersn/}{Nicholas M. Boers}, Staff Software Engineer\cr
  Jobber, Edmonton AB\cr
  Last modified \code{Rd: 2023-10-30}
}
\seealso{
  \code{\link{importPolys}}, \code{\link{importEvents}}, \code{\link{importGSHHS}}
}
\keyword{file}
